Bin Animator Update!
Fixed a major bug of the bin animator icon changer not functioning whatsoever.

Fixed the log file not working whatsoever. Now it fully functions.

Added the "stop" command in the command terminal for those that's running it by terminal

Added another requirement for an image file

Overall just cleaned up the code.