Update README.md
Made an entire setup file for those that want it to run in the background on their system. This would include installing and uninstalling it from the background processes.

The README has been fully redesigned. to encourage those to run it as a background process because it's the most efficiently

Temporarily removed, all printing to terminal. Hopefully it can return it eventually.